requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
python-datauri==0.2.8
six==1.11.0
Werkzeug==0.14.1
